+++ plugins/Passbolt/DirectorySync/tests/TestCase/Actions/GroupUserSyncActionTest.php
@@ -555,6 +555,60 @@ class GroupUserSyncActionTest extends DirectorySyncIntegrationTestCase
     }
 
     /**
+     * PB-1431: Fix: LDAP: a notification should not be sent to a group administrator requesting him to add a non-active user.
+     *
+     * Scenario: a non active groupUser has been added to a ldap group, not yet added in Passbolt.
+     * But the passbolt group has access to shared passwords.
+     * Expected result: No email notification should be sent. An ignore report should be broadcasted.
+     *
+     * @group DirectorySync
+     * @group DirectorySyncGroupUser
+     * @group DirectorySyncGroupUserAdd
+     */
+    public function testDirectorySyncGroupUser_PB1431()
+    {
+        // Init AppShellBootstrap to handle email notifications.
+        AppShellBootstrap::init();
+
+        $defaultGroupAdmin = 'edith@passbolt.com';
+        $this->setDefaultGroupAdminUser($defaultGroupAdmin);
+        $defaultGroupAdmin = $this->Users->findByUsername($defaultGroupAdmin)->first();
+        $this->initAction();
+
+        $userEntry = $this->mockDirectoryEntryUser(['fname' => 'ruth', 'lname' => 'ruth', 'foreign_key' => UuidFactory::uuid('user.id.ruth')]);
+        $this->mockDirectoryEntryGroup('accounting');
+        $this->mockDirectoryUserData('ruth', 'ruth', 'ruth@passbolt.com');
+        $this->mockDirectoryGroupData('accounting', [
+            'group_users' => [
+                $userEntry->directory_name,
+            ],
+        ]);
+
+        $reports = $this->action->execute();
+
+        $this->assertEquals(count($reports), 1);
+        $expectedUserGroupReport = [
+            'action' => Alias::ACTION_CREATE,
+            'model' => Alias::MODEL_GROUPS_USERS,
+            'status' => Alias::STATUS_IGNORE,
+            'type' => Alias::MODEL_GROUPS,
+            'message' => 'The user ruth@passbolt.com could not be added to the group Accounting because he has not yet activated his account.',
+        ];
+        $this->assertReport($reports[0], $expectedUserGroupReport);
+
+        // Group user for default admin should not exist.
+        $this->assertGroupUserNotExist(null, ['group_id' => UuidFactory::uuid('group.id.accounting'), 'user_id' => $defaultGroupAdmin->id]);
+
+        // Frances should not be in group users and directoryRelations.
+        $this->assertGroupUserNotExist(null, ['group_id' => UuidFactory::uuid('group.id.marketing'), 'user_id' => UuidFactory::uuid('user.id.ruth')]);
+        $this->assertDirectoryRelationEmpty();
+
+        // No email notification should have been sent to the group manager.
+        $this->get('/seleniumtests/showLastEmail/ada@passbolt.com');
+        $this->assertResponseCode(500);
+    }
+
+    /**
      * Scenario: A groupUser has been added to a group in ldap and already exist in passbolt
      * Expected result: do nothing. ignore report.
      *